Combine "udp_file_win.cc" and "udp_file_posix.cc".
